requirements.txt
Babel==2.3.4
DAWG-Python==0.7.2
Django==1.9.5
django-appconf==1.0.1
django-ckeditor==5.0.3
django-debug-toolbar==1.4
django-filter==0.13.0
django-geoip==0.5.2
django-mptt==0.8.3
django-mptt-urls==2.0.3
django-phonenumber-field==1.1.0
django-registration==2.1.2
django-smart-selects==1.2.2
django-treenav==1.0.0
docopt==0.6.2
phonenumberslite==7.4.4
Pillow==3.2.0
psycopg2==2.6.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.393658.3725883
python-slugify==1.2.0
pytz==2016.6.1
requests==2.9.1
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.1.19
Unidecode==0.4.19
xlrd==0.9.4
